Move onerror to helpers.py
